Extremely hacky goto implementation
Created label search built into compile , have to recompile with known labels in main in order to get working.
Short version: Doesn't need to change the cpu emulator to get the cpu assembler to implement labels since it compiles to raw 
address, labels are preserved but only as 'checkpoints' that are basically the equivalent of a NOP.

Uses a boolean firstrun = false vs firstrun = true this really is not how it should be done but is the only *simple* way.
Alot of verbose checking goes on within the get_bytes function for compiling the opcode itself but oh well.
Goto injection now works and as confirmed by hex editing it now allows replacement of label names with addresses so the goto should succeed. Soon it should be possible to easily implement the following opcodes: JMP LABEL (which is just a goto anyways) and SUB (LABEL) as well as JSR (LABEL) which is basically just goto reimplemented with the command to put onto the return stack.
So yay! (and the virtual machine cpu emulator doesn't have to change for this change)